requirements.txt
pysel-lang~=0.0.2
orjson~=3.8.0
